Squashed README and CONTRIBUTING
Rationale: There was a "Contributing" section
in the README. Why not just put contribution
information right there, for everyone to see
as part of the larger document with all the
rest of the information about the project as
well? This seems like a good idea for the
consolidation of documentation.